+++ acl/acl_test.go
@@ -151,6 +151,14 @@ func (m *mock) SetPCControl(controller uint32, enable bool) (bool, error) {
 	return true, nil
 }
 
+func (m *mock) GetAntiPassback(controller uint32) (types.AntiPassback, error) {
+	return types.Disabled, fmt.Errorf("NOT IMPLEMENTED")
+}
+
+func (m *mock) SetAntiPassback(controller uint32, antipassback types.AntiPassback) (bool, error) {
+	return false, fmt.Errorf("NOT IMPLEMENTED")
+}
+
 func (m *mock) RestoreDefaultParameters(controller uint32) (bool, error) {
 	return true, nil
 }